Title: 특별한 문자, Time: 10ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240308/특별한 문자/special-character.cpp
@@ -9,7 +9,7 @@ int main() {
     string strs;
     cin >> strs;
 
-    unordered_map<char, int> umap;
+    unordered_map<char, long long> umap;
 
     for(int i=0; i<strs.size(); i++)
     {
@@ -18,7 +18,7 @@ int main() {
 
     int now = INT_MAX;
     char first = '0';
-    unordered_map<char, int>::iterator iter;
+    unordered_map<char, long long>::iterator iter;
     for(iter=umap.begin(); iter!=umap.end(); iter++)
     {
         if(iter->second < now)
